Rename JavaProject-LeagueTables-ClassLeague to JavaProject-LeagueTables-LeagueClass
This Java project deals with football league tables. An example of a typical football league table is shown below:

Scottish League Division 1
P	 W 	D 	L 	F 	A 	Pt
Gretna 				16 	10 	3 	3 	43 	20 	33
Partick Thistle 			16 	8 	4 	4 	25 	29 	28
Hamilton Acas 			16	 7	 5 	4 	19 	20 	26
St Johnstone 			16	 6 	6 	4 	20 	16 	24
Livingstone 			16 	 6 	6 	4 	21 	15 	24
Dundee 				16 	7 	2 	7 	21 	18 	23
Clyde 				16 	5 	7 	4 	21 	17 	22
Ross County 			16 	4 	4 	8 	14 	24 	16
Queen of South 			16 	3 	3 	10 	11 	31 	12
Airdrie United 			16 	3 	2 	11 	14 	25 	11

For each team, the table shows the number of games Played (W) and, of those, the number Won (W), Drawn (D) and Lost (L); then the number of goals scored FOR the team (F) and AGAINST the team (A), and finally the Points (Pt).

My Java code samples represent a league table like the one above by means of two classes, Team and League.

Notice that the Team class has six instance variables (name, won, drawn, lost, goalsFor and goalsAgainst) and has setter methods for all except name. There is no setter for name because it is set once and for all by the constructor when a Team object is created. In addition, there are getter methods for all the instance variables, and also methods for returning the number of games played, the goal difference and the points, all of which can be calculated from the values of the instance variables.

The League class declares two instance variables, name and teams (an array of Team objects). It has a constructor, which sets the league’s name to reference the constructor's String argument, and sets teams to reference an array whose size is the second argument of the constructor. This constructor is declared to be private. This means that it is not possible to create a new League object by invoking the constructor directly from the OUWorkspace. Instead, The League Class will have a class method called loadLeague() whose purpose is to create a new League object (using the constructor) and fill it with teams whose details are contained in a text file.