Create LeagueTables|TeamClass
This is the second file of the Java project dealing with football league tables. This file contains my Java code for the Team class, which represents an array of teams within the football league itself.

The Team class code is written in a manner that arranges the teams in a League object in the usual descending order (highest number of points at the top). This is accomplished through a public instance method called sort(), which is embedded within the League class. While writing the code for this class, I found out that I could easily accomplish my goal of putting the teams in their winning descending order if I first defined a natural order for objects of the class Team. Ideally, I wanted the teams to be arranged in a descending order of points scored; and if two teams have the same points, then I wanted them to be in a descending order of their goals difference. If two teams tied on both points and goals difference, then no more ordering was required. To arrange for the ordering, I had to alter the header for the Team class and write a suitable compareTo() method (without overriding the equals() or hash() methods).

To test my compareTo() method, I had to write the code for the sort() method in my League class. This required only a single statement that uses a method from one of the 'collection utility' classes to arrange the Team objects in teams in the order described above. 

Executing the following code (through Java’s Main Method or through the command line):

League myLeague = League.loadLeague();
myLeague.sort();
myLeague.display();

should show the teams in the correct order.